Move resource/cmd under cmd/juju/resource
