+++ cmd/juju/resource/export_test.go
@@ -1,7 +1,14 @@
-package cmd
+// Copyright 2016 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package resource
+
+import (
+	"github.com/juju/juju/cmd/modelcmd"
+)
 
 func ListCharmResourcesCommandChannel(c *ListCharmResourcesCommand) string {
-	return c.channel
+	return modelcmd.InnerCommand(c).(*ListCharmResourcesCommand).channel
 }
 
 func ShowServiceCommandTarget(c *ShowServiceCommand) string {
@@ -19,3 +26,7 @@ func UploadCommandService(c *UploadCommand) string {
 }
 
 var FormatServiceResources = formatServiceResources
+
+func SetResourceLister(c modelcmd.Command, lister ResourceLister) {
+	modelcmd.InnerCommand(c).(*ListCharmResourcesCommand).resourceLister = lister
+}
