+++ cmd/juju/resource/list_charm_resources_test.go
@@ -1,7 +1,7 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package cmd_test
+package resource_test
 
 import (
 	"strings"
@@ -14,7 +14,7 @@ import (
 	charmresource "gopkg.in/juju/charm.v6-unstable/resource"
 
 	"github.com/juju/juju/charmstore"
-	resourcecmd "github.com/juju/juju/resource/cmd"
+	resourcecmd "github.com/juju/juju/cmd/juju/resource"
 )
 
 var _ = gc.Suite(&ListCharmSuite{})
@@ -70,7 +70,7 @@ func (s *ListCharmSuite) TestOkay(c *gc.C) {
 	s.client.ReturnListResources = [][]charmresource.Resource{resources}
 
 	command := resourcecmd.NewListCharmResourcesCommand()
-	command.ResourceLister = s.client
+	resourcecmd.SetResourceLister(command, s.client)
 	code, stdout, stderr := runCmd(c, command, "cs:a-charm")
 	c.Check(code, gc.Equals, 0)
 
@@ -96,7 +96,7 @@ func (s *ListCharmSuite) TestNoResources(c *gc.C) {
 	s.client.ReturnListResources = [][]charmresource.Resource{{}}
 
 	command := resourcecmd.NewListCharmResourcesCommand()
-	command.ResourceLister = s.client
+	resourcecmd.SetResourceLister(command, s.client)
 	code, stdout, stderr := runCmd(c, command, "cs:a-charm")
 	c.Check(code, gc.Equals, 0)
 
@@ -168,7 +168,7 @@ music     1
 	for format, expected := range formats {
 		c.Logf("checking format %q", format)
 		command := resourcecmd.NewListCharmResourcesCommand()
-		command.ResourceLister = s.client
+		resourcecmd.SetResourceLister(command, s.client)
 		args := []string{
 			"--format", format,
 			"cs:a-charm",
@@ -192,7 +192,7 @@ func (s *ListCharmSuite) TestChannelFlag(c *gc.C) {
 	}
 	s.client.ReturnListResources = [][]charmresource.Resource{resources}
 	command := resourcecmd.NewListCharmResourcesCommand()
-	command.ResourceLister = s.client
+	resourcecmd.SetResourceLister(command, s.client)
 
 	code, _, stderr := runCmd(c, command,
 		"--channel", "development",
