+++ cmd/juju/resource/show_service.go
@@ -1,7 +1,7 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package cmd
+package resource
 
 import (
 	"github.com/juju/cmd"
