+++ cmd/juju/resource/show_service_test.go
@@ -1,7 +1,7 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package cmd_test
+package resource_test
 
 import (
 	"time"
@@ -14,8 +14,8 @@ import (
 	charmresource "gopkg.in/juju/charm.v6-unstable/resource"
 	"gopkg.in/juju/names.v2"
 
+	resourcecmd "github.com/juju/juju/cmd/juju/resource"
 	"github.com/juju/juju/resource"
-	resourcecmd "github.com/juju/juju/resource/cmd"
 )
 
 var _ = gc.Suite(&ShowServiceSuite{})
